aither-adk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aither-adk"
version = "0.1.0a1"
description = "AitherOS Agent Development Kit — Build AI agents that work with any LLM backend"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Aitherium", email = "hello@aitherium.com" },
]
keywords = ["ai", "agents", "llm", "aitheros", "ollama", "openai", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
openai = ["openai>=1.0.0"]
all = ["anthropic>=0.18.0", "openai>=1.0.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.20.0",
]

[project.urls]
Homepage = "https://aitherium.com"
Repository = "https://github.com/Aitherium/AitherOS-Alpha"
Documentation = "https://github.com/Aitherium/AitherOS-Alpha#readme"

[project.scripts]
aither-serve = "adk.server:main"
aither-bug = "adk.bugreport:main"

[tool.hatch.build.targets.wheel]
packages = ["adk"]

[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
